Separate tracing middleware from error middleware
